FIX: Prevent PHP error if attempting to delete non existent alternative file via api_delete_alternative_file(). [t35455]
git-svn-id: http://svn.resourcespace.com/svn/rs/trunk@24167 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ include/resource_functions.php
@@ -5016,10 +5016,23 @@ function add_alternative_file($resource,$name,$description="",$file_name="",$fil
     return sql_insert_id();
     }
 
+/**
+ * delete_alternative_file
+ *
+ * @param   int   $resource   Resource id.
+ * @param   int   $ref        Alternative file id.
+ * 
+ * @return  bool   false on error else true.
+ */
 function delete_alternative_file($resource,$ref)
     {
     # Delete any uploaded file.
     $info=get_alternative_file($resource,$ref);
+    if (!$info)
+        {
+        return false;
+        }
+
     $path=get_resource_path($resource, true, "", true, $info["file_extension"], -1, 1, false, "", $ref);
     hook('delete_alternative_file_extra', '', array($path));
     if (file_exists($path)) {unlink($path);}
